fix teacher entity - not created

+++ src/main/java/com/fpt/onlineTest/model/CoursesRegistration.java
@@ -15,7 +15,7 @@ import java.io.Serializable;
 public class CoursesRegistration implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @ManyToOne
     @JoinColumn(name = "student_id")
@@ -28,5 +28,5 @@ public class CoursesRegistration implements Serializable {
     @ManyToOne
     @JoinColumn(name = "teacher_id")
     private Teacher teacher;
-//
+
 }
